+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added resolving of tenant access payload for MCP users (via the header `X-Sre-Mcp-User-Name`)
 - Added support for K8S event-driven
 - Added two new vendors for K8S `SRE_K8S_AGENT` and `SRE_K8S_WATCHER` (see `docs/event_driven.md` for more details)
-- Added new endpoint `PUT /platforms/k8s/{platform_id}` to enable/disable K8S event-driven
+- Added new endpoint `PATCH /platforms/k8s/{platform_id}` to enable/disable K8S event-driven
 
 ### Fixed
 - Fixed on-prem API not passing query params (e.g. `customer_id`) for non-GET requests (DELETE, PUT, etc.)